+++ README.md
@@ -1,2 +1,7 @@
 # Domi_Nations
 Student project game
+
+## Installation
+* Importing jar to project
+File -> Project Structure -> Modules -> + -> Library java
+
