+++ src/main/java/com/isep/domination/Jeu.java
@@ -1,3 +1,5 @@
+package com.isep.domination;
+
 import java.awt.Color;
 import java.awt.Font;
 import java.util.ArrayList;
