+++ src/main/java/com/isep/domination/Pioche.java
@@ -1,3 +1,5 @@
+package com.isep.domination;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
