+++ src/main/java/com/isep/domination/Plateau.java
@@ -1,3 +1,5 @@
+package com.isep.domination;
+
 import edu.princeton.cs.introcs.StdDraw;
 
 public class Plateau {
@@ -57,7 +59,7 @@ public class Plateau {
 		double x = StdDraw.mouseX();
 		double y = StdDraw.mouseY();
 		
-		//test cases joueur 1 -- validé faire la même pour les autres joueurs 2,3,4
+		//test cases joueur 1 -- validï¿½ faire la mï¿½me pour les autres joueurs 2,3,4
 		for (int i=0; i<5; i++){
 			for (int j=0; j<5; j++){
 				if((x >= 5 + i*TAILLE_CASE + i*TAILLE_LIGNE && x<= 5 + (i+1)*TAILLE_CASE + (i+1)*TAILLE_LIGNE) && (y <= 715 - j*TAILLE_CASE - j*TAILLE_LIGNE && y >= 715 - (j+1)*TAILLE_CASE - (j+1)*TAILLE_LIGNE)){
